+++ pkg/host/host_openbsd.go
@@ -14,5 +14,4 @@ func isSupported(c *prog.Syscall, sandbox string) (bool, string) {
 func init() {
 	checkFeature[FeatureCoverage] = unconditionallyEnabled
 	checkFeature[FeatureNetworkInjection] = unconditionallyEnabled
-	checkFeature[FeatureSandboxSetuid] = unconditionallyEnabled
 }
